types: split types.go to multiple files
The types.go file became to big and made it harder to find specific
functionality.
Split the files into smaller, more specific, ones.
